move useUserSearchResults.ts into ScheduleMeetingModal
